+++ bp/orgalogs/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, include
 
-from bp.views import OrgaLogView
+from .orga.views import OrgaLogView
 
 orgalog_patterns = [
     path('<pk>/', OrgaLogView.as_view(), name='orga_log_detail'),
